+++ src/main/java/chatroom/controllers/DangKy.java
@@ -11,23 +11,30 @@ import javax.websocket.Session;
 import javax.websocket.server.ServerEndpoint;
 import org.bson.Document;
 import com.mongodb.client.FindIterable;
-import chatroom.Config.DataBase;
 import chatroom.models.*;
 
 @ServerEndpoint(value = "/dang-ky", encoders = { ResponseEncoder.class }, decoders = { UserDecoder.class })
 public class DangKy {
 	@OnMessage
 	public void onMessage(User user, Session session) throws IOException, EncodeException {
-		DataBase db = new DataBase();
+		Database db = new Database();
 		db.init();
 		HashMap<String,Object> account = new HashMap<String,Object>();
 		FindIterable<Document> result = db.findOne("username", user.getUsername());
+		Response response = new Response();
 		if(result.first() != null) {
-			ResponseSender.response(session, "error", 1, "Tên tài khoản đã được sử dụng");
+			response.setStatus("error");
+			response.setCode(1);
+			response.setMessage("Tên tài khoản đã được sử dụng");
+			ResponseSender.send(session, response);
 			return;
 		}
 		else {
-			ResponseSender.response(session, "success", 0, "Đăng ký thành công!");
+			response.setStatus("success");
+			response.setCode(0);
+			response.setMessage("Đăng ký thành công!");
+			ResponseSender.send(session, response);
+
 			account.put("token", user.getToken());
 			account.put("createAt", user.getDatetime());
 			account.put("password", user.getPassword());
